requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

albumentations==1.3.0
numpy==1.23.5
Pillow==9.3.0
#torch==1.13.1+cu116
#torchvision==0.14.1+cu116
opencv-python==4.6.0.66
matplotlib==3.6.3
pandas==1.5.2
pytorch-lightning==1.9.0
scikit-learn==1.2.0
tqdm==4.64.1
